Esqueleto del util
