+++ src/windowsappication/WindowsAppicationReto2.java
@@ -3,42 +3,37 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package windowsappication.reto2;
+package windowsappication;
 
+import com.sun.istack.internal.logging.Logger;
+import java.io.IOException;
 import javafx.application.Application;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
+import windowsappication.controller.MainWindowController;
 
 /**
  *
  * @author Gaizka Andres
  */
 public class WindowsAppicationReto2 extends Application {
-    
+
     @Override
-    public void start(Stage primaryStage) {
-        Button btn = new Button();
-        btn.setText("Say 'Hello World'");
-        btn.setOnAction(new EventHandler<ActionEvent>() {
-            
-            @Override
-            public void handle(ActionEvent event) {
-                System.out.println("Hello World!");
-            }
-        });
+    public void start(Stage stage) throws Exception {
+        FXMLLoader loader = new FXMLLoader(getClass().
+            getResource(
+            "/windowsclientappication/reto2/view/MainWindowController.fxml"));
+        Parent root = (Parent) loader.load();
+        MainWindowController mainWindowController = loader.getController();
+        mainWindowController.setStage(stage);
+        mainWindowController.initStage(root);
         
-        StackPane root = new StackPane();
-        root.getChildren().add(btn);
-        
-        Scene scene = new Scene(root, 300, 250);
-        
-        primaryStage.setTitle("Hello World!");
-        primaryStage.setScene(scene);
-        primaryStage.show();
     }
 
     /**
