+++ src/windowsappication/controller/AdminWindowController.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package windowsappication.reto2.controller;
+package windowsappication.controller;
 
 /**
  *
  * @author Gaizka Andres
  */
-public class CreditCardWindowController {
+public class AdminWindowController {
     
 }
