+++ src/windowsappication/controller/HelpWindowController.java
@@ -1,0 +1,25 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package windowsappication.controller;
+
+import javafx.scene.Parent;
+import javafx.stage.Stage;
+
+/**
+ *
+ * @author Gaizka Andres
+ */
+public class HelpWindowController {
+
+    void setStage(Stage stage) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    void initStage(Parent root) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+    
+}
