+++ src/windowsappication/controller/LoginWindowController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package windowsappication.reto2.controller;
+package windowsappication.controller;
 import java.io.IOException;
 import java.util.Optional;
 import java.util.logging.Logger;
