+++ src/windowsappication/controller/UploadDocWindowController.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package windowsappication.reto2.controller;
+package windowsappication.controller;
 
 /**
  *
  * @author Gaizka Andres
  */
-public class InfoDocWindowController {
+public class UploadDocWindowController {
     
 }
