+++ src/windowsappication/view/VerDocumento.fxml
